espanol/requirements.txt
pandas==2.0.0
numpy==1.24.0
matplotlib==3.7.1
seaborn==0.12.2
scikit-learn==1.2.2
statsmodels==0.14.0
pymysql==1.1.0
python-dotenv==1.0.0
jupyter==1.0.0
notebook==7.0.0
